requirements.txt
# Core ML / NLP
#torch>=2.0.0
#transformers>=4.38.0
#sentence-transformers>=2.6.0
#scikit-learn>=1.4.0
#numpy>=1.26.0
#pandas>=2.2.0

# Resume parsing
#PyMuPDF>=1.23.0          # PDF parsing (fitz)
#python-docx>=1.1.0       # DOCX parsing
#pdfplumber>=0.11.0       # PDF text extraction backup

# Text processing
#nltk>=3.8.1
#spacy>=3.7.0
#regex>=2024.1.0

# Generative AI (resume rewriting)0
#openai>=1.12.0           # optional fallback
#accelerate>=0.27.0       # for local model inference

# ATS Scoring
#rouge-score>=0.1.2
#bert-score>=0.3.13

# Web UI
#streamlit>=1.32.0
#streamlit-extras>=0.4.0

# Utilities
#python-dotenv>=1.0.0
#tqdm>=4.66.0
#loguru>=0.7.2
streamlit>=1.33.0
sentence-transformers>=2.7.0
transformers>=4.41.0
torch>=2.2.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
python-dotenv>=1.0.1
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
python-docx>=1.1.0
loguru>=0.7.2
